+++ CHANGELOG.md
@@ -7,3 +7,13 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## 0.1.0 - 05/01/2022
 
 - Initial release
+
+## 0.1.1 - 10/01/2023
+
+- Second release
+
+## 0.1.2 - 04/01/2023
+
+- Workspace request
+- Data source explorer
+